refactor(taskflow): extract feature.account
Move AccountsModel, AuthFlowModel, AuthMode, SessionModel, account actions
(EditProfile, StartLogin, LoginRequested, AccountLoggedIn, LoginFailed,
LoadAccountsRequested/Succeeded/Failed, SwitchAccount, LogoutAccount),
accountsReducer, authFlowReducer, sessionReducer, LoginScreen, SwitcherScreen,
ProfileScreen, AccountRow, and AccountSwitchTest into
…feature.account. EditProfile remains cross-cutting: collaboratorsReducer
in reducer.AccountReducers imports it from the new package.
