+++ examples/taskflow/composeApp/src/commonTest/kotlin/org/reduxkotlin/sample/taskflow/reducer/RootReducersTest.kt
@@ -2,28 +2,30 @@ package org.reduxkotlin.sample.taskflow.reducer
 
 import kotlinx.collections.immutable.persistentListOf
 import kotlinx.collections.immutable.persistentMapOf
-import org.reduxkotlin.sample.taskflow.action.AccountLoggedIn
-import org.reduxkotlin.sample.taskflow.action.EditProfile
-import org.reduxkotlin.sample.taskflow.action.LoadAccountsSucceeded
-import org.reduxkotlin.sample.taskflow.action.LoginFailed
-import org.reduxkotlin.sample.taskflow.action.LoginRequested
-import org.reduxkotlin.sample.taskflow.action.LogoutAccount
 import org.reduxkotlin.sample.taskflow.action.Refresh
-import org.reduxkotlin.sample.taskflow.action.StartLogin
-import org.reduxkotlin.sample.taskflow.action.SwitchAccount
 import org.reduxkotlin.sample.taskflow.core.AccountId
 import org.reduxkotlin.sample.taskflow.core.AccountSummary
 import org.reduxkotlin.sample.taskflow.core.AppSettingsModel
 import org.reduxkotlin.sample.taskflow.core.Theme
+import org.reduxkotlin.sample.taskflow.feature.account.AccountLoggedIn
+import org.reduxkotlin.sample.taskflow.feature.account.AccountsModel
+import org.reduxkotlin.sample.taskflow.feature.account.AuthFlowModel
+import org.reduxkotlin.sample.taskflow.feature.account.AuthMode
+import org.reduxkotlin.sample.taskflow.feature.account.EditProfile
+import org.reduxkotlin.sample.taskflow.feature.account.LoadAccountsSucceeded
+import org.reduxkotlin.sample.taskflow.feature.account.LoginFailed
+import org.reduxkotlin.sample.taskflow.feature.account.LoginRequested
+import org.reduxkotlin.sample.taskflow.feature.account.LogoutAccount
+import org.reduxkotlin.sample.taskflow.feature.account.StartLogin
+import org.reduxkotlin.sample.taskflow.feature.account.SwitchAccount
+import org.reduxkotlin.sample.taskflow.feature.account.accountsReducer
+import org.reduxkotlin.sample.taskflow.feature.account.authFlowReducer
 import org.reduxkotlin.sample.taskflow.feature.settings.SetBotEnabled
 import org.reduxkotlin.sample.taskflow.feature.settings.SetFailureRate
 import org.reduxkotlin.sample.taskflow.feature.settings.SetLatency
 import org.reduxkotlin.sample.taskflow.feature.settings.SetOnline
 import org.reduxkotlin.sample.taskflow.feature.settings.SetTheme
 import org.reduxkotlin.sample.taskflow.feature.settings.appSettingsReducer
-import org.reduxkotlin.sample.taskflow.model.AccountsModel
-import org.reduxkotlin.sample.taskflow.model.AuthFlowModel
-import org.reduxkotlin.sample.taskflow.model.AuthMode
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertNull
